requirements.txt
# Core dependencies
gymnasium>=0.29.0
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
matplotlib>=3.7.0
tensorboard>=2.13.0

# Optional dependencies for advanced features
plotly>=5.15.0
wandb>=0.15.0
streamlit>=1.25.0
gradio>=3.40.0

# Development dependencies
black>=23.0.0
ruff>=0.0.280
pytest>=7.4.0
pytest-cov>=4.1.0
mypy>=1.5.0

# Environment-specific dependencies
cartpole-env>=0.1.0  # For CartPole environment
